[delete] remove BucketsMultiTable and DatabaseCompactMode
