--- paimon-flink/paimon-flink-common/src/main/java/org/apache/paimon/flink/action/DatabaseCompactMode.java
@@ -1,52 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.paimon.flink.action;
-
-import javax.annotation.Nullable;
-
-import java.io.Serializable;
-
-/**
- * There are two modes for database compact.
- *
- * <p>1) DIVIDED mode, start a sink for each table, the compaction of the new table requires
- * restarting the job.
- *
- * <p>2) COMBINED mode, start a single combined sink for all tables, the new table will be
- * automatically compacted.
- */
-public enum DatabaseCompactMode implements Serializable {
-    DIVIDED,
-    COMBINED;
-
-    public static DatabaseCompactMode fromString(@Nullable String mode) {
-        if (mode == null) {
-            return DIVIDED;
-        }
-
-        switch (mode.toLowerCase()) {
-            case "divided":
-                return DIVIDED;
-            case "combined":
-                return COMBINED;
-            default:
-                throw new UnsupportedOperationException("Unsupported mode: " + mode);
-        }
-    }
-}
